GridSear/grid_walk for testing
GridSearch is now based on astropy tables as input and output. grid_walk internally still works with Parameter objects, but that’s not visible to the "outside". Please see the comments in the code.

Comments on stopping criteria:
Generic: if the number of current_models is >= n_max_mods, no new models will be created (can be tested by setting n_max_models = 1 in the config file). Note that for #models < n_max_mods new models are still added. If we want to avoid this, how do we select the models to add?
Iteration count, delta_chi2: not yet implemented, needs iteration counter column in AllModels.table

Can be tested by executing lines 4-43 in test_import.py.

Please test…