Add Executor class to handle different modes of execution e.g. Local, Slurm etc.
The executor type is specified in the config_file --> 
executable_settings. Like with all "type" attribtes, this corresponds to 
the name of a class, to be found in executables.py.

The executor needs to be instantiated and passed to the ModelIterator 
object. This is currently done in the script (test_iterator.py) but can 
probably be done elsewhere as to be hidden from the user.

TODO: add slurm commands.