Rename plotbin->plotbin4dyn
Several users had clashes between their locally installed versions of 
Capellari's plotbin and the version we packaged up inside DYNAMITE. We 
(currently) can't use the latest version of plotbin, as it it seemingly 
incompatible with one of the imported Schwpy routines. So a workaround 
for now is to rename our packaged-up version to something else, 
plotbin4dyn. This should only be a temportary solution - longterm let's 
re-write the plotting routine.