Rename plotbin->plotbin4dyn (#23)
* Rename plotbin->plotbin4dyn

Several users had clashes between their locally installed versions of 
Capellari's plotbin and the version we packaged up inside DYNAMITE. We 
(currently) can't use the latest version of plotbin, as it it seemingly 
incompatible with one of the imported Schwpy routines. So a workaround 
for now is to rename our packaged-up version to something else, 
plotbin4dyn. This should only be a temportary solution - longterm let's 
re-write the plotting routine.

* Fixed import statement in plotbin4dyn/cap_display_pixels.py

Co-authored-by: Prashin Jethwa <pjethwa@ma023782.ads.eso.org>
Co-authored-by: Chia-vie <christine.ackerl@univie.ac.at>