Fix the "stars" component bug (#46)
* Fix the "stars" component bug:
By referencing the stellar component via system.cmp_list[2] we rely on the stellar component listed as the third component in the configuration file. This will break if the configuration file has a different sequence of components or more/less components in the future so that the third is not the "stars" component any more. By selecting the component with the name "stars" the behaviour is more robust now.

* Take advantage of existing System.get_component_from_name(...) method.

Co-authored-by: Thomas I. Maindl <thomas.maindl@sdb.ltd>