Allow NULL dyn.kinematics.GaussHermite object + write method to add PSF (#41)
Conversion from legacy kinematics to ECSV format

* Allow NULL dyn.kinematics.GaussHermite object + write method to add PSF

Sabine was unable to convert legacy kinematic file into the ECSV file.
To do this requires (i) creating an NULL dyn.kinematics.GaussHermite
object (i.e. instantiated with no arguments) and then using the it's
convert_to_old_format.

But there was an error on creating an NULL dyn.kinematics.GaussHermite
object. To fix this, fix big on line 13 of data.py and add check on line
40.

Next, we realised that the converted kinematics file did not have the
PSF in the header. So added the method Integrated.add_psf_to_datafile
to allow this.

Note that the specific layout of the kinemtic data-file now looks
different (e.g. no !!omap, and fewer lines). This may be due to an
Astropy version change..? But things still work.

Other minor changes:
i) fix spelling of filenames 'ecvs' --> 'ecsv' (filnames and also in
test/*config.yaml)
ii) add plotbin4dyn to the gitignore

* Remove hardcoded n_gh=4 from GaussHermite methods 'read_file_old_format' and 'convert_to_old_format'

* Included ecvs --> ecsv change to docs/tutorials

* Added comment in config files: input files of test galaxy are the same

* Updated format of the old kinematics file after convertion

Co-authored-by: Prashin Jethwa <pjethwa@ma023782.ads.eso.org>
Co-authored-by: sthater <sthater87@web.de>