Redirected the legacy_fortran output (both stdout and stderr) to .log files and added logging debug messages pointing to these output files. Setting the logging level to logging.WARNING or above should allow a silent Dynamite run now.
Note that it is the responsibility of the application that uses Dynamite to set up logging. In the two tests we demonstrate an elaborate setup involving logging to the console and a file (test_nnls.py) and a quasi-minimal setup (test_orbit_losvds.py). Also note that the Python kernel needs to be restarted between runs to reset logging.