Proper logging and controlling output quantity (#65)
* First shot at implementing logging with the logging module. The philosophy is that the main application is responsible for setting up logging (destinations, level of detail, formatting) and Dynamite contributes.

This first implementation sets up logging in test_nnls.py (debug+ messages to test_nnls.log and info+ messages to the console) and config_reader.py creates a log.

Note how each method that uses logging grabs a logger (logger = logging.getLogger(f'{__name__}.{__class__.__name__}')).

* Some tweaking and formatting

* This is a in-between commit, don't pull just yet ;-)

* Redirected the legacy_fortran output (both stdout and stderr) to .log files and added logging debug messages pointing to these output files.
Setting the logging level to logging.WARNING or above should allow a silent Dynamite run now.

Note that it is the responsibility of the application that uses Dynamite to set up logging. In the two tests we demonstrate an elaborate setup involving logging to the console and a file (test_nnls.py) and a quasi-minimal setup (test_orbit_losvds.py). Also note that the Python kernel needs to be restarted between runs to reset logging.

* Decluttered config_reader log output by demoting INFO to DEBUG log entries.

* Added method dynamite_logging(), a "Dynamite default" logging configuration to the config reader. It is called when Configuration is instantiated with the reset_logging=True argument and should not be used if logging has been configured outside of Dynamite already.

* Moved logger definition in the __init()__ methods. Classes that have logging now use the self.logger....(...) methods.

* Changed Python 3.6 -> 3.7 in setup

Co-authored-by: Thomas I. Maindl <thomas.maindl@sdb.ltd>