Allow default settings for histogram quantities (#103)
Defaults can be used by specifying `default` in the confil file, or 
ommitting these entries. Class attribute setting is handled by three 
methods `set_default_hist_{width/center/bins}` which are called in the 
`init` method of `Kinematics`. Specific implementations of these methods 
are in specific kinematics sub-classes. For `GaussHermite` the defaults 
are:
- width = 2 * max(|v| + 3*sig)
- center = 0 
- nbins so that (approximately) dv = 0.1 * min(sig)
For `BayesLOSVD` the defaults are read directly from the source 
datafile.

This required making some additional changes to:
- config reader: cast hist settings to numeric types only if the are not 
'default'
- config reader: if `LegacyWeightSolver` is used, then all histogram 
settings must be the same. Implementation of this changed slightly.
- in order to be able to instantiate empty kinemtic classes, some minor 
changes to `data.py` and passing the `type` kwarg in the 
`super().__init__`