Implement Quenneville+21 mirror fix (#189)
* Implement Quenneville+21 mirror fix

Quenneville+21 claim that our orbit mirroring is incorrect and suggest a 
fix. This is implemented in the new `orblib_f_new_mirror.f90` which has 
been edited in lines 817-825.

To use this, specify:
```
orblib_settings:
   use_new_mirroring : True
```
in the congfiguration file. If set to False, or this entry is omitted, 
then the old mirroring is used.

* Flip signs in qgrid_store

Forgotten first time. Thanks Giulia for pointing this out :)

* Update default and docs

Set default value of `use_new_mirroring` to True, and update 
documentation.

* Last minute change: added orblib_new_mirror to the distclean section in Makefile.

Co-authored-by: Thomas I. Maindl <thomas.maindl@sdb.ltd>